requirements.txt
Django==6.0
Pillow==12.0.0
django-ckeditor==6.7.3
django-crispy-forms==2.5
crispy-bootstrap5==2025.6
django-widget-tweaks==1.5.0
python-decouple==3.8
gunicorn==20.1.0
psycopg2-binary==2.9.6
